+++ main.go
@@ -3,12 +3,17 @@ package main
 import (
 	"fmt"
 	"log"
+	"log/syslog"
 	"os"
 	"sync"
 )
 
+var (
+	ErrLogger = log.New(os.Stderr, "log-shuttle: ", log.LstdFlags)
+)
+
 const (
-	VERSION = "0.5.5"
+	VERSION = "0.6.0"
 )
 
 func MakeBasicBits(config ShuttleConfig) (reader *Reader, stats chan NamedValue, drops, lost *Counter, logs chan LogLine, deliverableBatches chan *Batch, programStats *ProgramStats, bWaiter, oWaiter *sync.WaitGroup) {
@@ -37,8 +42,16 @@ func Shutdown(deliverableLogs chan LogLine, stats chan NamedValue, deliverableBa
 
 func main() {
 	var config ShuttleConfig
+	var err error
 
 	config.ParseFlags()
+
+	if config.LogToSyslog {
+		ErrLogger, err = syslog.NewLogger(syslog.LOG_ERR|syslog.LOG_SYSLOG, 0)
+		if err != nil {
+			log.Fatalf("Unable to setup syslog logger: %s\n", err)
+		}
+	}
 
 	if config.PrintVersion {
 		fmt.Println(VERSION)
@@ -46,7 +59,7 @@ func main() {
 	}
 
 	if !config.UseStdin() {
-		log.Fatalln("No stdin detected.")
+		ErrLogger.Fatalln("No stdin detected.")
 	}
 
 	reader, stats, _, _, logs, deliverableBatches, _, batchWaiter, outletWaiter := MakeBasicBits(config)